mysite/requirements.txt
Django>=4.2.0
gunicorn>=21.2.0
whitenoise>=6.0.0
django-environ>=0.11.2
psycopg2-binary>=2.9.9
pillow>=10.0.0
django-tinymce>=3.6.1
python-dotenv
dj-database-url
django-static-fontawesome
django-crispy-forms
crispy-bootstrap5

# Для бэкапов и восстановления
django-dbbackup